Log "info" when newsletter not found (#28088)
Some requests for newsletter forms are made for newsletters that do not
exist. This is currently being logged as an error, but that suggests
it's a problem with the platform that we should fix. In reality, a
missing newsletter is a content issue rather than a problem with the
platform, so it should be logged as a piece of information rather than
an error.

Co-authored-by: David Blatcher <30567854+dblatcher@users.noreply.github.com>