Log "info" when tag index not found (#28096)
Requests to retrieve tag index pages result in a 404 (not found) when
the tag index does not exist. This is currently being logged as an
error, but that suggests it's a problem with the platform that we should
fix. In reality, a tag index that does not exist is not something that's
likely to change, or that we can "fix", so it should be logged as a
piece of information rather than an error.