Remove `directToRouter` troubleshooter test
The admin app contains a troubleshooting page for testing dotcom URLs.
This included a test for checking whether router instances for a
particular URL can be hit. However, it appears this test is no longer
working.

Removing it allows us to remove a dependency on the AWS EC2 SDK v1,
which would otherwise need to be updated to v2.

Co-authored-by: Roberto Tyley <52038+rtyley@users.noreply.github.com>