Use `onComplete` tp handle Future result after `putMetricsData` operation
Previously, the Future was wrapped by a Try but that would not catch the Future failure. Try synchronously captures exceptions thrown in the scope of its block: https://www.scala-lang.org/api/2.13.6/scala/util/Try.html. We tested using invalid credentials in the AWS client and the Try was not capturing the error.

See more:
https://medium.com/%40shantanukamath/why-try-scala-error-handling-d71d2bd44d8c
https://medium.com/@lashford.disney/error-handling-pitfalls-in-scala-2c1a18bebfc8

Co-authored-by: Marjan Kalanaki <marjan.kalanaki@guardian.co.uk>