Create non-AWS group
We would like to test whether we can add a new group and have it work
alongside the list of centrally configured groups.

Whether or not this works we will revert it, as on frontend we don't
want a single group for all non-AWS dependencies. This is just for
testing.

This replaces the previous AWS group, as this was not successful in
overriding the centrally configured AWS group.