Unfinished refactor: try adding `blocks` field to `ArticlePage` etc
Jamie made a great point when reviewing https://github.com/guardian/frontend/pull/28437,
which introduced `BlocksOn[Page]` to pair together `page` & `blocks`:

https://github.com/guardian/frontend/pull/28437#pullrequestreview-3535449581

> I wonder looking at this whether `blocks` should be a property on one of the page types, since blocks are part of the page?

...this commit tries out that approach, which looked really great - but in the end, there
was a small wrinkle that I'd need more thought to evaluate before adopting that model:

We have several page classes that use blocks - `ArticlePage`, `InteractivePage`,
`GalleryPage` - and when those are constructed, that's generally alongside the
full _range_ of blocks in that content. However, for `LiveBlogPage`, it will be
a smaller - variable - range of blocks, depending on what `LiveBlogCurrentPage`
is being viewed.

In some ways `BlocksOn[Page]` is superior, because it doesn't imply that
the blocks we see are _all_ the blocks for that page.