Refactor: Introduce `BlocksOn[Page]` parameter object
This work aims to support the Live-Harness endpoint functionality
being introduced by Fred in https://github.com/guardian/frontend/pull/28417 -
we can reduce the size of that implementation by taking a few small refactors
in `ArticleController` & `InteractiveController`. Those refactors simplify the
code in those controllers, as well as supporting the forthcoming code in the
new `LiveHarnessController`.

## Introduce `BlocksOn[Page]` parameter object

`page` and `blocks` are two parameters that are passed together in
several method calls, suggesting that they belong together in a single
parameter object:

https://refactoring.com/catalog/introduceParameterObject.html
https://refactoring.guru/introduce-parameter-object
https://docondev.com/blog/2020/6/2/refactoring-introduce-parameter-object

As the job of the `LiveHarnessController` is to accept new interactive atoms,
and then use them to transform an existing page-&-block combination, the transformation
can be expressed a little more clearly when a single `pageBlocks` value encompasses
both page & blocks.

## Controller method refactors

On `ArticleController`:

* Three identical methods (`renderArticle()`, `renderJson()`, `renderEmail()`)
  are now implemented only in `renderArticle()`, with the other 2 methods
  delegating to `renderArticle()`.
* Common code in `renderItem()` & `renderArticle()` has been extracted into
  `mapAndRender()`, which also supports an optional `pageBlocks` modifier,
  useful for the forthcoming live-harness work.

On `InteractiveController`:

* `toModel()` is changed to `modelAndRender()`, which now incorporates common
  `Either`-handling logic (previously in `renderItem()`), and accepts a
  rendering-function parameter, invoked on the happy-path when the Page
  has been found & modelled. This enables us to concisely express the rendering
  we want to perform, useful for the forthcoming live-harness work.

On `LiveBlogController`:

* `match` statements within `renderEmail()` & `renderJson()` have been simplified -
  they were previously matching on tuple-pairs where the second value did not
  affect the match.